+++ ext/QuantumCliffordHeckeExt/QuantumCliffordHeckeExt.jl
@@ -4,10 +4,14 @@ using QECCore
 import QECCore: code_n, code_s, code_k, rate, distance
 using DocStringExtensions
 
-import QuantumClifford, LinearAlgebra
+import QuantumClifford
+import QuantumClifford: gf2_row_echelon_with_pivots!
+import LinearAlgebra
 import Hecke: Group, GroupElem, AdditiveGroupElem,
     GroupAlgebra, GroupAlgebraElem, FqFieldElem, representation_matrix, dim, base_ring,
-    multiplication_table, coefficients, abelian_group, group_algebra, rand, gens
+    multiplication_table, coefficients, abelian_group, group_algebra, rand, gens, order,
+    is_commutative, FqPolyRingElem, residue_ring, coeff, zero_matrix, mod1, lift, ZZ, gen,
+    matrix, ncols, nrows, degree
 import Nemo
 import Nemo: characteristic, matrix_repr, GF, ZZ, lift
 
@@ -20,11 +24,12 @@ import QECCore: AbstractECC, CSS,
 
 # exported from extension so that Documenter.jl sees them when autogenerating API lists
 export hgp, two_block_group_algebra_codes, generalized_bicycle_codes, bicycle_codes, haah_cubic_codes,
-    LPCode, LiftedCode, honeycomb_color_codes
+    LPCode, LiftedCode, honeycomb_color_codes, LaCross
 
 include("util.jl")
 include("types.jl")
 include("lifted.jl")
+include("lacross.jl")
 include("lifted_product.jl")
 
 end # module