+++ lib/QECCore/src/QECCore.jl
@@ -3,6 +3,8 @@ module QECCore
 using SparseArrays
 using LinearAlgebra
 using Combinatorics
+using Graphs
+using Random: GLOBAL_RNG, AbstractRNG, randperm, rand
 
 using DocStringExtensions
 
@@ -14,10 +16,15 @@ export AbstractECC, AbstractQECC, AbstractCECC, AbstractCSSCode, AbstractDistanc
 export Perfect5, Cleve8, Gottesman
 
 # CSS Codes
-export Toric, Bitflip3, Phaseflip3, Shor9, Steane7, Surface, CSS, QuantumReedMuller, Triangular488, Triangular666
+export Toric, Bitflip3, Phaseflip3, Shor9, Steane7, Surface, CSS, QuantumReedMuller, Triangular488, Triangular666,
+DelfosseReichardt, DelfosseReichardtRepCode, DelfosseReichardt823, QuantumTannerGraphProduct, CyclicQuantumTannerGraphProduct,
+TillichZemor, random_TillichZemor_code
 
 # Classical Codes
 export RepCode, ReedMuller, RecursiveReedMuller, Golay, Hamming
+
+# utilities
+export search_self_orthogonal_rm_codes
 
 include("interface.jl")
 include("codes/util.jl")
@@ -38,11 +45,18 @@ include("codes/quantum/surface.jl")
 include("codes/quantum/bitflipcode.jl")
 include("codes/quantum/gottesman.jl")
 include("codes/quantum/color_codes.jl")
+include("codes/quantum/quantumtannergraphproduct.jl")
+include("codes/quantum/tillichzemor.jl")
 
 # Reed-Muller Codes
 include("codes/classical/reedmuller.jl")
 include("codes/classical/recursivereedmuller.jl")
 include("codes/quantum/quantumreedmuller.jl")
+
+# Delfosse-Reichardt Codes
+include("codes/quantum/delfosse_reichardt_code.jl")
+include("codes/quantum/delfosse_reichardt_repcode.jl")
+include("codes/quantum/delfosse_reichardt_823_code.jl")
 
 function __init__()
     if isdefined(Base.Experimental, :register_error_hint)